use NSInteger and such; add some stuff for typestatus 1.1.1

+++ BulletinBoard/BBSectionInfo.h
@@ -1,8 +1,8 @@
 @interface BBSectionInfo : NSObject
 
-+ (instancetype)defaultSectionInfoForType:(unsigned)type;
++ (instancetype)defaultSectionInfoForType:(NSUInteger)type;
 
-@property unsigned notificationCenterLimit;
+@property NSUInteger notificationCenterLimit;
 @property (nonatomic, retain) NSString *sectionID;
 
 @end
